+++ src/plugin/Cat_Plugin.h
@@ -1,0 +1,10 @@
+#ifndef SHOW__CAT_PLUGIN_H
+#define SHOW__CAT_PLUGIN_H
+
+
+#include "Plugin.h"
+
+extern Plugin Cat_Plugin;
+
+
+#endif