+++ src/plugin/Plugin.h
@@ -10,6 +10,7 @@
 
 
 typedef struct {
+    List options;
     const char* (*get_description)();
     const char* (*get_name)();
     int (*run)(int fd_in, int argc, char* argv[], List options, Error* error);